Fix Linux + Windows CI build errors on integration (round 4)
Both the Linux (gcc/clang -Werror) and Windows (MSVC /WX) CI jobs fail on merged
feature code. This round fixes both platforms:

Windows MSVC (/WX, warnings-as-errors):
- HouseInteriorMap::ReplaceWallWithDoorway / UpdateRoomComponentTextures: comment
  the unreferenced newRoomGuid / room parameter names (C4100).
- Housing::OnQuestCompleted: remove the unused previousLevel local (C4189).

Linux gcc/clang:
- ChallengeMode.cpp: include DatabaseEnv.h for the global CharacterDatabase.
- Garrison.cpp: cast FOLLOWER_TYPE_GARRISON to uint8 in the follower-type
  ternary (-Werror=extra enumerated/non-enumerated conditional).
- ChallengeModeHandler.cpp: mark the currently-unwired RollMythicPlusRewardItem
  helper [[maybe_unused]] (-Wunused-function).

+++ src/server/game/Garrison/Garrison.cpp
@@ -1932,7 +1932,7 @@ GarrisonError Garrison::ClaimMissionReward(uint32 missionRecID)
                 WorldPackets::Garrison::GarrisonFollower oldFollowerState = follower->PacketInfo;
 
                 GarrFollowerEntry const* followerEntry = sGarrFollowerStore.LookupEntry(follower->PacketInfo.GarrFollowerID);
-                uint8 followerTypeID = followerEntry ? followerEntry->GarrFollowerTypeID : FOLLOWER_TYPE_GARRISON;
+                uint8 followerTypeID = followerEntry ? followerEntry->GarrFollowerTypeID : uint8(FOLLOWER_TYPE_GARRISON);
 
                 follower->PacketInfo.Xp += followerXP;
